+++ packages/ladle/lib/shared/types.ts
@@ -189,6 +189,7 @@ export type Config = {
   outDir: string;
   base?: string;
   mode?: string;
+  expandStoryTree?: boolean;
   noWatch: boolean;
   hotkeys: {
     fullscreen: string[];
@@ -280,3 +281,26 @@ export type GetUserViteConfig = {
 export type EntryData = {
   [key: string]: ParsedStoriesResult;
 };
+
+export type MetaJson<M extends KnownMeta = KnownMeta> = {
+  about: {
+    homepage: string;
+    github: string;
+    version: number;
+  };
+  stories: { [key: string]: MetaJsonStory<M> };
+};
+
+export type MetaJsonStory<M extends KnownMeta = KnownMeta> = {
+  name: string;
+  levels: string[];
+  meta: M;
+  locStart: number;
+  locEnd: number;
+};
+
+export interface KnownMeta {
+  iframed?: boolean;
+  width?: string | number | "xsmall" | "small" | "medium" | "large";
+  mockDate?: string;
+}