+++ e2e/css/CHANGELOG.md
@@ -1,5 +1,33 @@
 # test-css
 
+## 0.0.93
+
+### Patch Changes
+
+- Updated dependencies [[`189cd5a`](https://github.com/tajo/ladle/commit/189cd5a21f13856a25e01010f0ee3f4a0a1a1a29), [`ce4d509`](https://github.com/tajo/ladle/commit/ce4d5093694239860f3e3df5b4b67f69580e5cb9)]:
+  - @ladle/react@5.0.2
+
+## 0.0.92
+
+### Patch Changes
+
+- Updated dependencies [[`cdb6117`](https://github.com/tajo/ladle/commit/cdb6117c8060fd2b6b4b07f4f61c844f88447fda)]:
+  - @ladle/react@5.0.1
+
+## 0.0.91
+
+### Patch Changes
+
+- Updated dependencies [[`7930e38`](https://github.com/tajo/ladle/commit/7930e38647dc9ef40dd92ff578815f77acb50ef6), [`1178c37`](https://github.com/tajo/ladle/commit/1178c37bfeb5aba0a3249811151f463beacdecef), [`eaac9e9`](https://github.com/tajo/ladle/commit/eaac9e9a1e2c33b19bae02b58ee01657198d513b), [`a4859ce`](https://github.com/tajo/ladle/commit/a4859cea3cb49dd93d32736fed5c556bcdf53418)]:
+  - @ladle/react@5.0.0
+
+## 0.0.90
+
+### Patch Changes
+
+- Updated dependencies [[`5af1fea`](https://github.com/tajo/ladle/commit/5af1fea5ac61adaf2a60eced125be6e47fac492b)]:
+  - @ladle/react@4.1.2
+
 ## 0.0.89
 
 ### Patch Changes